+++ test/unit/Validation/ValidNumberTest.php
@@ -1,0 +1,44 @@
+<?php
+/**
+ * @author    Andrew Coulton <andrew@ingenerator.com>
+ * @licence   proprietary
+ */
+
+namespace test\unit\Ingenerator\PHPUtils\Validation;
+
+
+use Ingenerator\PHPUtils\Validation\ValidNumber;
+
+class ValidNumberTest extends \PHPUnit_Framework_TestCase
+{
+
+    /**
+     * @testWith ["", false]
+     *           ["junk", false]
+     *           ["0", false]
+     *           ["12", false]
+     *           ["13", true]
+     *           [18, true]
+     */
+    public function test_it_validates_minimum_as_greater_or_equal($value, $expect)
+    {
+        $this->assertSame($expect, ValidNumber::minimum($value, 13));
+    }
+
+    /**
+     * @expectedException \InvalidArgumentException
+     */
+    public function test_it_throws_for_unknown_rule_name()
+    {
+        ValidNumber::rule('random nonsense');
+    }
+
+    public function test_it_returns_validator_name()
+    {
+        $this->assertSame(
+            'Ingenerator\PHPUtils\Validation\ValidNumber::minimum',
+            ValidNumber::rule('minimum')
+        );
+    }
+
+}