+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### Unreleased
 
+* Extract the basic ValidNumber class for validating minimum
+
 ### v0.1.2 (2018-02-13)
 
 * Fix signature of InvalidUserDateTime::createFromMutable for PHP >= 5.6 compatibility